Move DateTimeUtil to quasar.time
