+++ yggdrasil/src/main/scala/quasar/yggdrasil/table/cf/Util.scala
@@ -18,8 +18,9 @@ package quasar.yggdrasil.table.cf
 
 import quasar.blueeyes._
 import quasar.contrib.std.errorImpossible
-import quasar.precog._
+import quasar.precog.BitSet
 import quasar.precog.common._
+import quasar.time.DateTimeUtil
 import quasar.yggdrasil._
 import quasar.yggdrasil.table._
